chore(suite-native): Mobile trade: move countriesWhitelist to trading-consts
