+++ suite-native/trading-atoms/src/index.ts
@@ -34,4 +34,3 @@ export * from './utils/general/cryptoIdUtils';
 export * from './utils/general/currencyUtils';
 export * from './utils/general/receiveAccountUtils';
 export * from './utils/general/tradeableAssetUtils';
-export * from './utils/residence/countriesWhitelist';
