+++ suite-native/trading-residence/package.json
@@ -23,6 +23,7 @@
         "@suite-native/intl": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/trading-atoms": "workspace:*",
+        "@suite-native/trading-consts": "workspace:*",
         "@suite-native/trading-state": "workspace:*",
         "@trezor/styles": "workspace:*",
         "expo-localization": "~17.0.7",
