+++ suite-native/trading-residence/tsconfig.json
@@ -13,6 +13,7 @@
         { "path": "../intl" },
         { "path": "../navigation" },
         { "path": "../trading-atoms" },
+        { "path": "../trading-consts" },
         { "path": "../trading-state" },
         { "path": "../../packages/styles" },
         { "path": "../test-utils" }
